+++ Orar/OrarApp/SubjectTableModel.cpp
@@ -74,7 +74,9 @@ void SubjectTableModel::PopulateModel(QString aName)
 
 void SubjectTableModel::ClearContent()
 {
-    beginRemoveRows(QModelIndex(), 0, mContext.GetSubjectSize() - 1);//emit signal to notify view that a new row is removed
-    mContext.DeleteSubjects();
-    endRemoveRows();
+    if (mContext.GetSubjectSize() > 0) {
+        beginRemoveRows(QModelIndex(), 0, mContext.GetSubjectSize()-1);//emit signal to notify view that a new row is removed
+        mContext.DeleteSubjects();
+        endRemoveRows();
+    }
 }
